# library-book-network-jwt-app/library-book-network-backend/src/main/java/org/aadi/lbn/feedback/FeedbackResponse.java
package org.aadi.lbn.feedback;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}